move test options to root of tests directory
